# Conflicts:
#	backofisbackend/src/main/java/com/omniteam/backofisbackend/controller/OrdersController.java
#	backofisbackend/src/main/java/com/omniteam/backofisbackend/service/OrderService.java
#	backofisbackend/src/main/java/com/omniteam/backofisbackend/service/implementation/OrderServiceImpl.java

+++ backofisbackend/src/main/java/com/omniteam/backofisbackend/entity/OrderDetail.java
@@ -19,7 +19,7 @@ public class OrderDetail extends BaseEntity  {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "order_detail_id")
-    private Long orderDetailId;
+    private Integer orderDetailId;
 
 
     @Column(name="status")
